# openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Minimarket Plus API",
    "description": "API REST del sistema backend Minimarket Plus. Gestiona productos, categorías, carritos, inventario, ventas y usuarios con roles (cliente, cajero, administrador). Todos los endpoints de negocio requieren autenticación HTTP Basic; use el botón Authorize (credenciales de demostración en el README).",
    "contact": {
      "name": "Equipo Minimarket Plus - PBY2202",
      "email": "karla.santigut@gmail.com"
    },
    "license": {
      "name": "Uso académico - Duoc UC"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Entorno local de desarrollo"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/api/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener un producto por ID",
        "description": "Busca un producto específico por su identificador único.",
        "operationId": "obtenerProductoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un producto con el ID indicado"
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar un producto existente",
        "description": "Reemplaza los datos del producto identificado por el ID.",
        "operationId": "actualizarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del producto a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Nuevos datos del producto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              },
              "examples": {
                "productoActualizado": {
                  "description": "productoActualizado",
                  "value": {
                    "nombre": "Coca-Cola 1.5L",
                    "precio": 1890,
                    "stock": 24,
                    "categoria": {
                      "id": 1
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un producto con el ID indicado"
          },
          "200": {
            "description": "Producto actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "summary": "Eliminar un producto",
        "description": "Elimina definitivamente el producto identificado por el ID.",
        "operationId": "eliminarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del producto a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado correctamente"
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un producto con el ID indicado"
          }
        }
      }
    },
...
    "/api/productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Listar todos los productos",
        "description": "Retorna el catálogo completo de productos registrados en el minimarket.",
        "operationId": "listarProductos",
        "responses": {
          "200": {
            "description": "Listado obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Producto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "summary": "Crear un nuevo producto",
        "description": "Registra un producto en el catálogo. La categoría referenciada debe existir previamente.",
        "operationId": "guardarProducto",
        "requestBody": {
          "description": "Datos del producto a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              },
              "examples": {
                "nuevoProducto": {
                  "description": "nuevoProducto",
                  "value": {
                    "nombre": "Coca-Cola 1.5L",
                    "precio": 1890,
                    "stock": 24,
                    "categoria": {
                      "id": 1
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "201": {
            "description": "Producto creado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único de la categoría",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre único de la categoría",
            "example": "Bebidas"
          }
        },
        "description": "Categoría para clasificar los productos del minimarket"
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del producto",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre comercial del producto",
            "example": "Coca-Cola 1.5L"
          },
          "precio": {
            "type": "number",
            "description": "Precio unitario en pesos chilenos",
            "format": "double",
            "example": 1890
          },
          "stock": {
            "type": "integer",
            "description": "Unidades disponibles en bodega",
            "format": "int32",
            "example": 24
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        },
        "description": "Producto del catálogo del minimarket"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Autenticación HTTP Basic contra los usuarios registrados en H2",
        "name": "basicAuth",
        "scheme": "basic"
      }
    }
  }
}